Added Patika deleted and updated

+++ patikaKlonu/src/com/patikadev/View/OperatorGUI.java
@@ -6,8 +6,8 @@ import com.patikadev.Model.User;
 import javax.swing.*;
 import javax.swing.event.TableModelEvent;
 import javax.swing.table.DefaultTableModel;
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
+import java.awt.*;
+import java.awt.event.*;
 import java.util.ArrayList;
 
 public class OperatorGUI extends JFrame {
@@ -110,6 +110,29 @@ public class OperatorGUI extends JFrame {
         patikaMenu.add(updateMenu);
         patikaMenu.add(deleteMenu);
 
+        updateMenu.addActionListener(e -> {
+            int select_id = Integer.parseInt(tbl_patikaList.getValueAt(tbl_patikaList.getSelectedRow(),0).toString());
+            UpdatePatikaGUI updatePatikaGUI = new UpdatePatikaGUI(Patika.getFetch(select_id));
+            updatePatikaGUI.addWindowListener(new WindowAdapter() {
+                @Override
+                public void windowClosed(WindowEvent e) {
+                    loadPatikaModel();
+                }
+            });
+        });
+
+        deleteMenu.addActionListener(e -> {
+            if (Helper.confirm("sure")){
+                int select_id = Integer.parseInt(tbl_patikaList.getValueAt(tbl_patikaList.getSelectedRow(),0).toString());
+                if (Patika.delete(select_id)){
+                    Helper.showMessage("success");
+                    loadPatikaModel();
+                }else{
+                    Helper.showMessage("error");
+                }
+            }
+        });
+
         mdl_patikaList = new DefaultTableModel();
         Object[] col_patikaList = {"ID", "Patika Adı"};
         mdl_patikaList.setColumnIdentifiers(col_patikaList);
@@ -117,8 +140,18 @@ public class OperatorGUI extends JFrame {
         loadPatikaModel();
 
         tbl_patikaList.setModel(mdl_patikaList);
+        tbl_patikaList.setComponentPopupMenu(patikaMenu);
         tbl_patikaList.getTableHeader().setReorderingAllowed(false);
         tbl_patikaList.getColumnModel().getColumn(0).setMaxWidth(75);
+
+         tbl_patikaList.addMouseListener(new MouseAdapter() {
+             @Override
+             public void mousePressed(MouseEvent e) {
+                 Point point = e.getPoint();
+                 int selected_row = tbl_patikaList.rowAtPoint(point);
+                 tbl_patikaList.setRowSelectionInterval(selected_row,selected_row);
+             }
+         });
         // ###
 
         //Kişi Ekleme Butonu
@@ -147,12 +180,14 @@ public class OperatorGUI extends JFrame {
             if (Helper.isFieldEmpty(fld_userId)){
                 Helper.showMessage("fill");
             }else {
-                int userId = Integer.parseInt(fld_userId.getText());
-                if (User.deleteFunction(userId)){
-                    Helper.showMessage("success");
-                    loadUserModel();
-                }else {
-                    Helper.showMessage("error");
+                if (Helper.confirm("sure")){
+                    int userId = Integer.parseInt(fld_userId.getText());
+                    if (User.deleteFunction(userId)){
+                        Helper.showMessage("success");
+                        loadUserModel();
+                    }else {
+                        Helper.showMessage("error");
+                    }
                 }
             }
         });
